requirements.txt
langgraph>=0.4.7
langchain>=0.3.20
langchain-core>=0.3.15
langchain-tavily>=0.2.0
langgraph-supervisor
python-dotenv>=1.0.0 